Comments for future generations.
Also, I've been running this for about an hour and it just found
its first solution (the known solution).  Hooray!  We'll see if
there are any more.  I expect it will take about 3 days to run on
my laptop, and I don't care enough to optimize it to run faster.